202304_PowerBI_Stars & Orgs MA Data
Added two files from Jan and Feb 2023. One is a pdf of PowerBI dashboard that'd allow users sort Medicare Advantage (MA) contracts by state, Star rating, and type of organization. The second document is a PowerBI dashboard that shows the number of MA contracts owned by each of the 313 MA plan parent organizations that existed as of December 2022.